Rename Fireblocks CW unified module
