+++ src/providers/fireblocks/cw/core/fireblocks-cw-unified.module.ts
@@ -1,0 +1,39 @@
+import { Module } from '@nestjs/common';
+import { FireblocksCoreModule } from './fireblocks-core.module';
+import { FireblocksCwUnifiedService } from './fireblocks-cw-unified.service';
+import { AdminAuditModule } from './modules/admin-audit.module';
+import { AdminDestinationsModule } from './modules/admin-destinations.module';
+import { AdminGasOperationsModule } from './modules/admin-gas-operations.module';
+import { AdminSecurityModule } from './modules/admin-security.module';
+import { AdminWithdrawalsModule } from './modules/admin-withdrawals.module';
+import { CwDepositModule } from './modules/cw-deposit.module';
+import { CwPortfolioModule } from './modules/cw-portfolio.module';
+import { CwTransactionsModule } from './modules/cw-transactions.module';
+import { CwTransfersModule } from './modules/cw-transfers.module';
+
+const ADMIN_MODULES = [
+  AdminSecurityModule,
+  AdminDestinationsModule,
+  AdminWithdrawalsModule,
+  AdminGasOperationsModule,
+  AdminAuditModule,
+];
+
+const NON_ADMIN_MODULES = [
+  CwDepositModule,
+  CwPortfolioModule,
+  CwTransfersModule,
+  CwTransactionsModule,
+];
+
+@Module({
+  imports: [FireblocksCoreModule, ...ADMIN_MODULES, ...NON_ADMIN_MODULES],
+  providers: [FireblocksCwUnifiedService],
+  exports: [
+    FireblocksCoreModule,
+    FireblocksCwUnifiedService,
+    ...ADMIN_MODULES,
+    ...NON_ADMIN_MODULES,
+  ],
+})
+export class FireblocksCwUnifiedModule {}
