+++ src/providers/fireblocks/cw/core/modules/admin-destinations.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { FireblocksCoreModule } from '../../core/fireblocks-core.module';
-import { AdminDestinationsService } from './admin-destinations.service';
+import { FireblocksCoreModule } from '../fireblocks-core.module';
+import { AdminDestinationsService } from '../services/admin-destinations.service';
 
 @Module({
   imports: [FireblocksCoreModule],
