+++ src/providers/fireblocks/cw/core/modules/admin-gas-operations.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { FireblocksCoreModule } from '../../core/fireblocks-core.module';
-import { AdminGasOperationsService } from './admin-gas-operations.service';
+import { FireblocksCoreModule } from '../fireblocks-core.module';
+import { AdminGasOperationsService } from '../services/admin-gas-operations.service';
 
 @Module({
   imports: [FireblocksCoreModule],
