+++ src/providers/fireblocks/cw/core/modules/cw-portfolio.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { FireblocksCoreModule } from '../core/fireblocks-core.module';
-import { CwPortfolioService } from './cw-portfolio.service';
+import { FireblocksCoreModule } from '../fireblocks-core.module';
+import { CwPortfolioService } from '../services/cw-portfolio.service';
 
 @Module({
   imports: [FireblocksCoreModule],
