+++ src/providers/fireblocks/cw/core/modules/cw-transfers.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { FireblocksCoreModule } from '../core/fireblocks-core.module';
-import { CwTransfersService } from './cw-transfers.service';
+import { FireblocksCoreModule } from '../fireblocks-core.module';
+import { CwTransfersService } from '../services/cw-transfers.service';
 
 @Module({
   imports: [FireblocksCoreModule],
