+++ src/providers/fireblocks/cw/core/services/cw-portfolio.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { FireblocksClientProvider } from '../core/providers/fireblocks-client.provider';
+import { FireblocksClientProvider } from '../providers/fireblocks-client.provider';
 
 @Injectable()
 export class CwPortfolioService {
