+++ src/providers/fireblocks/cw/core/services/cw-transfers.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { FireblocksClientProvider } from '../core/providers/fireblocks-client.provider';
-import { FireblocksErrorMapper } from '../core/providers/fireblocks-error-mapper';
-import { FireblocksResilience } from '../core/providers/fireblocks-resilience';
+import { FireblocksClientProvider } from '../providers/fireblocks-client.provider';
+import { FireblocksErrorMapper } from '../providers/fireblocks-error-mapper';
+import { FireblocksResilience } from '../providers/fireblocks-resilience';
 
 export interface TransferCommand {
   source: string;
